Adds fish shell support.
This uses an unsightly hack where the environment variables are spit out
as eval-able fish commands.

This has the advantage of only needing to care about the environment
variables that are set by the environment, and not the implementation.

It has the disadvantage where it is fiddly; a stray misquoted character
can wreak havoc.